+++ CHANGELOG.md
@@ -29,9 +29,13 @@
 - New #164: Add `SuggestedPostApproved` type.
 - New #164: Add `suggestedPostApproved` field to `Message` type.
 - New #164: Add `SuggestedPostApprovalFailed` type.
+- New #164: Add `suggestedPostApprovalFailed` field to `Message` type.
 - New #164: Add `SuggestedPostDeclined` type.
+- New #164: Add `suggestedPostDeclined` field to `Message` type.
 - New #164: Add `SuggestedPostPaid` type.
+- New #164: Add `suggestedPostPaid` field to `Message` type.
 - New #164: Add `SuggestedPostRefunded` type.
+- New #164: Add `suggestedPostRefunded` field to `Message` type.
 
 ## 0.9.1 July 4, 2025